Rename containers to pages in upload plugin
