+++ packages/core/admin/admin/src/components/PrivateRoute/index.js
@@ -32,7 +32,7 @@ const PrivateRoute = ({ component: Component, path, ...rest }) => (
 );
 
 PrivateRoute.propTypes = {
-  component: PropTypes.oneOfType([PropTypes.node, PropTypes.func, PropTypes.element]).isRequired,
+  component: PropTypes.any.isRequired,
   path: PropTypes.string.isRequired,
 };
