+++ packages/core/upload/admin/src/components/HomePageModalStepper/reducer.js
