+++ packages/core/upload/admin/src/hooks/useSelectTimestamps/index.js
@@ -1,5 +1,5 @@
 import { useSelector } from 'react-redux';
-import selectFileModelTimestamps from '../../containers/Initializer/selectors';
+import selectFileModelTimestamps from '../../components/Initializer/selectors';
 
 const useSelectTimestamps = () => {
   const timestamps = useSelector(selectFileModelTimestamps);
