+++ packages/core/upload/admin/src/index.js
@@ -9,11 +9,11 @@ import { CheckPagePermissions } from '@strapi/helper-plugin';
 import pluginPkg from '../../package.json';
 import pluginLogo from './assets/images/logo.svg';
 import pluginPermissions from './permissions';
-import App from './containers/App';
-import Initializer from './containers/Initializer';
-import SettingsPage from './containers/SettingsPage';
+import Initializer from './components/Initializer';
 import InputMedia from './components/InputMedia';
-import InputModalStepper from './containers/InputModalStepper';
+import InputModalStepper from './components/InputModalStepper';
+import App from './pages/App';
+import SettingsPage from './pages/SettingsPage';
 import reducers from './reducers';
 import trads from './translations';
 import pluginId from './pluginId';
