+++ packages/core/upload/admin/src/pages/HomePage/reducer.js
