+++ packages/core/upload/admin/src/pages/HomePage/utils/generateStringFromParams.js
