+++ packages/core/upload/admin/src/pages/SettingsPage/reducer.js
